docker-credential-gcr needed for push to GCR
I ran into a python error when trying to push to GCR without docker-credential-gcr installed. I propose that it be added to the prerequisites.
Another potential addition is `go get apimachinery` which I did not have installed prior to attempting to build cluster registry and caused my build to fail. I still am not sure what apimachinery is/does and why it would/wouldn't be installed anyway. I am happy to add this to the prerequisites too if it makes sense.